# src/main/java/com/heliosx/consultation_service/model/dto/request/ConsultationFormDto.java
package com.heliosx.consultation_service.model.dto.request;

import lombok.Builder;
import lombok.Getter;
import lombok.Setter;

import java.util.List;

@Getter
@Setter
@Builder
public class ConsultationFormDto {
    Long medicalConditionId;
    List<QuestionAnswerDto> questionAnswers;
}