# src/main/java/com/heliosx/consultation_service/model/dto/response/QuestionDto.java
package com.heliosx.consultation_service.model.dto.response;

import lombok.Getter;
import lombok.Setter;

@Getter
@Setter
public class QuestionDto {
    private Long id;
    private String questionText;
}